Initial commit, merging all repos in a single one.
There's quite a lot of waste to be dealt with yet, from past refactors
and cleanups. Some of the plugins are still just clones from the Sample
template (eg. the RedisSub isn't a thing yet, and the Run plugin is
probably broken in several ways).

As it stands now, as well, the project isn't in a proper shape for
CPAN-ification.. at the bare minimum, it needs another run through the
documentation, as well as adding tests all around.